Rename `rustc_mir_build/src/errors.rs` into `rustc_mir_build/src/diagnostics.rs`
