+++ compiler/rustc_mir_build/src/builder/mod.rs
@@ -44,7 +44,7 @@ use rustc_span::{Span, Symbol};
 
 use crate::builder::expr::as_place::PlaceBuilder;
 use crate::builder::scope::{DropKind, LintLevel};
-use crate::errors;
+use crate::diagnostics;
 
 pub(crate) fn closure_saved_names_of_captured_variables<'tcx>(
     tcx: TyCtxt<'tcx>,
@@ -931,7 +931,7 @@ impl<'a, 'tcx> Builder<'a, 'tcx> {
                 lint::builtin::UNREACHABLE_CODE,
                 lint_root,
                 target_loc.span,
-                errors::UnreachableDueToUninhabited {
+                diagnostics::UnreachableDueToUninhabited {
                     expr: target_loc.span,
                     orig: orig_span,
                     descr,
